Fixed Errors in Test
Ok so there were some typos present, but the main issues were that you cant import Paths as a module without specifying where Paths.py is located. The other that caused the runtime issue is the issue I was mentioning earlier in that singles is the only object that I am unable to open into a plot. To show the rest of the code works, coincidences were plotted instead.