Filter tests look good. Need to refine filter parameters
The Kalman filter is proving very useful. Still need to implement custom filter settings and testing out offsetting the data by the noise mean to see if the filter behaves even better.

Then need to see if we can find "true" energies to compare to and see if we get closer to actual energies by use of filters