Timestamp-based filter appears successful
After refinement of methods and reviewing on multiple sets of data, the filter appears successful in removing protons that the DAQ qualifies as protons but are most likely noise. Next step will be to first make the program more efficient; then attempt to somehow link the protons to the their most likley electron.